Former-commit-id: 4a07b886930ce3ff3038869de700741ce67c4e1c

+++ Sources/Internal/Render/Material/NMaterial.cpp
@@ -160,8 +160,8 @@ void NMaterial::BindParams(rhi::Packet& target)
         materialBufferBinding->lastValidPropertySemantic = NMaterialProperty::GetCurrentUpdateSemantic();
     }
 
-    target.vertexConstCount = activeVariantInstance->vertexConstBuffers.size();
-    target.fragmentConstCount = activeVariantInstance->fragmentConstBuffers.size();
+    target.vertexConstCount = static_cast<uint32>(activeVariantInstance->vertexConstBuffers.size());
+    target.fragmentConstCount = static_cast<uint32>(activeVariantInstance->fragmentConstBuffers.size());
     /*bind material const buffers*/
     for (size_t i = 0, sz = activeVariantInstance->vertexConstBuffers.size(); i < sz; ++i)
         target.vertexConst[i] = activeVariantInstance->vertexConstBuffers[i];
@@ -691,8 +691,8 @@ void NMaterial::RebuildTextureBindings()
         const rhi::ShaderSamplerList& fragmentSamplerList = currShader->GetFragmentSamplerList();
         const rhi::ShaderSamplerList& vertexSamplerList = currShader->GetVertexSamplerList();
 
-        textureDescr.fragmentTextureCount = fragmentSamplerList.size();
-        samplerDescr.fragmentSamplerCount = fragmentSamplerList.size();
+        textureDescr.fragmentTextureCount = static_cast<uint32>(fragmentSamplerList.size());
+        samplerDescr.fragmentSamplerCount = static_cast<uint32>(fragmentSamplerList.size());
         for (size_t i = 0, sz = textureDescr.fragmentTextureCount; i < sz; ++i)
         {
             RuntimeTextures::eDynamicTextureSemantic textureSemantic = RuntimeTextures::GetDynamicTextureSemanticByName(currShader->GetFragmentSamplerList()[i].uid);
@@ -720,8 +720,8 @@ void NMaterial::RebuildTextureBindings()
             DVASSERT(textureDescr.fragmentTexture[i].IsValid());
         }
 
-        textureDescr.vertexTextureCount = vertexSamplerList.size();
-        samplerDescr.vertexSamplerCount = vertexSamplerList.size();
+        textureDescr.vertexTextureCount = static_cast<uint32>(vertexSamplerList.size());
+        samplerDescr.vertexSamplerCount = static_cast<uint32>(vertexSamplerList.size());
         for (size_t i = 0, sz = textureDescr.vertexTextureCount; i < sz; ++i)
         {
             Texture* tex = GetEffectiveTexture(vertexSamplerList[i].uid);
